Slightly improving x86 functions (read: removing `_x86Add()`).
Javascript bitwise arithmetic clips to 32 bits. This means that we
don't need `& 0xffffffff` anywhere, or any other hacks to capture
overflows if we only need the first 32 bits.

But if we *do* care about overflow past 32 bits (as in the case of the
various x64 functions), we're forced to split up our work across
smaller integers, hence all the bitwise work over 16 bit ints, whose
overflow javascript won't clip.

The upshot of this is that `_x86Add()` is entirely unnecessary, and
`_x86Multiply()` doesn't require the final `&` step.